Removed __super support and clean up
The only reason for Base2.py and Base3.py was due to the existence
of the AxonType metaclass which was used to add a __super method.
That metaclass was always a mistake and __super was deprecated for
a long while.

This change removes the support for __super, and therefore removes
the need for Base2 and Base3. The examples that use super in some
very old code have been modified, and code in /Sketches also fixed
for the same reason (reasonable consistency)